# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: CDX Loap Huong Notification Service API
  description: |
    Comprehensive notification service API for CDX Loap Huong platform.

    ## Features
    - Device token management for push notifications
    - Notification history and management
    - User notification preferences
    - Admin broadcast capabilities
    - Retry mechanism for failed notifications
    - Real-time delivery tracking

    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.

    ## Rate Limiting
    - User endpoints: 1000 requests per minute
    - Admin endpoints: 100 requests per minute

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: CDX Development Team
    email: dev@cdx.com
  license:
    name: Proprietary
    url: https://cdx.com/license
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.cdx.com
    description: Production server
  - url: https://staging-api.cdx.com
    description: Staging server
security:
  - bearerAuth: []
paths:
...
  /device-tokens:
    get:
      tags:
        - Device Tokens
      summary: Get user device tokens
      description: Retrieve all device tokens for the authenticated user
      responses:
        '200':
          description: Device tokens retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Device Tokens
      summary: Register device token
      description: Register a new device token for push notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceTokenRequest'
            examples:
              ios_device:
                summary: iOS Device Token
                value:
                  token: 'ios_device_token_123456789'
                  platform: 'ios'
                  deviceName: 'iPhone 15 Pro'
                  osVersion: '17.0'
                  appVersion: '1.0.0'
              android_device:
                summary: Android Device Token
                value:
                  token: 'android_device_token_987654321'
                  platform: 'android'
                  deviceName: 'Samsung Galaxy S24'
                  osVersion: '14.0'
                  appVersion: '1.0.0'
      responses:
        '201':
          description: Device token registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          description: Device token already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /device-tokens/{token}:
    put:
      tags:
        - Device Tokens
      summary: Update device token
      description: Update an existing device token
      parameters:
        - name: token
          in: path
          required: true
          description: Device token to update
          schema:
            type: string
            example: 'ios_device_token_123456789'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeviceTokenRequest'
      responses:
        '200':
          description: Device token updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
        - Device Tokens
      summary: Delete device token
      description: Remove a device token
      parameters:
        - name: token
          in: path
          required: true
          description: Device token to delete
          schema:
            type: string
            example: 'ios_device_token_123456789'
      responses:
        '200':
          description: Device token deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Notification Management
...
  /admin/statistics:
    get:
      tags:
        - Admin
      summary: Get notification statistics
      description: Retrieve comprehensive notification statistics (Admin only)
      parameters:
        - name: period
          in: query
          description: Time period for statistics
          schema:
            type: string
            enum: [today, week, month, quarter, year, custom]
            default: week
        - name: startDate
          in: query
          description: Start date for custom period (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: End date for custom period (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/broadcast:
    post:
      tags:
        - Admin
      summary: Create broadcast notification
      description: Create a broadcast notification to multiple users (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BroadcastRequest'
            examples:
              immediate_broadcast:
                summary: Immediate Broadcast
                value:
                  title: 'System Maintenance'
                  body: 'Scheduled maintenance will occur tonight from 2-4 AM'
                  type: 'announcement'
                  priority: 'normal'
                  targetUsers: ['all']
                  channels: ['push', 'email']
              scheduled_broadcast:
                summary: Scheduled Broadcast
                value:
                  title: 'New Feature Release'
                  body: 'Check out our new features in the latest update'
                  type: 'announcement'
                  priority: 'normal'
                  targetUsers: ['all']
                  channels: ['push']
                  scheduledAt: '2024-01-15T10:00:00Z'
      responses:
        '201':
          description: Broadcast notification created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BroadcastResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication service
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  description: Error code
                details:
                  type: string
                  description: Detailed error message
                field:
                  type: string
                  description: Field that caused the error (for validation errors)
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              description: Response data

    # Health Schemas
...
    DeviceToken:
      type: object
      required:
        - id
        - token
        - platform
        - userId
        - createdAt
      properties:
        id:
          type: string
          description: Unique device token ID
        token:
          type: string
          description: Device token value
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        deviceName:
          type: string
          description: Device name
        osVersion:
          type: string
          description: Operating system version
        appVersion:
          type: string
          description: Application version
        userId:
          type: string
          description: User ID
        isActive:
          type: boolean
          description: Whether the token is active
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegisterDeviceTokenRequest:
      type: object
      required:
        - token
        - platform
      properties:
        token:
          type: string
          description: Device token value
          example: 'ios_device_token_123456789'
        platform:
          $ref: '#/components/schemas/DevicePlatform'
        deviceName:
          type: string
          description: Device name
          example: 'iPhone 15 Pro'
        osVersion:
          type: string
          description: Operating system version
          example: '17.0'
        appVersion:
          type: string
          description: Application version
          example: '1.0.0'
    UpdateDeviceTokenRequest:
      type: object
      properties:
        token:
          type: string
          description: New device token value
        deviceName:
          type: string
          description: Device name
        osVersion:
          type: string
          description: Operating system version
        appVersion:
          type: string
          description: Application version
    DeviceTokenResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/DeviceToken'
    DeviceTokenListResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/DeviceToken'

    # Notification Schemas
...
    Statistics:
      type: object
      properties:
        totalNotifications:
          type: integer
          description: Total number of notifications
        deliveredCount:
          type: integer
          description: Number of delivered notifications
        failedCount:
          type: integer
          description: Number of failed notifications
        pendingCount:
          type: integer
          description: Number of pending notifications
        successRate:
          type: number
          format: float
          description: Success rate (0-1)
        averageDeliveryTime:
          type: number
          format: float
          description: Average delivery time in milliseconds
        topNotificationTypes:
          type: array
          items:
            type: object
            properties:
              type:
                $ref: '#/components/schemas/NotificationType'
              count:
                type: integer
        topChannels:
          type: array
          items:
            type: object
            properties:
              channel:
                $ref: '#/components/schemas/NotificationChannel'
              count:
                type: integer
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
    StatisticsResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Statistics'
    BroadcastRequest:
      type: object
      required:
        - title
        - body
        - type
        - targetUsers
      properties:
        title:
          type: string
          description: Notification title
          example: 'System Maintenance'
        body:
          type: string
          description: Notification body
          example: 'Scheduled maintenance will occur tonight from 2-4 AM'
        type:
          $ref: '#/components/schemas/NotificationType'
        priority:
          $ref: '#/components/schemas/NotificationPriority'
          default: normal
        targetUsers:
          type: array
          items:
            type: string
          description: Target user IDs or 'all' for broadcast
          example: ['all']
        channels:
          type: array
          items:
            $ref: '#/components/schemas/NotificationChannel'
          description: Notification channels to use
          example: ['push', 'email']
        data:
          type: object
          description: Additional notification data
        scheduledAt:
          type: string
          format: date-time
          description: When to send the notification (optional)
    BroadcastResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - id
                - title
                - type
                - status
                - targetCount
              properties:
                id:
                  type: string
                  description: Broadcast notification ID
                title:
                  type: string
                type:
                  $ref: '#/components/schemas/NotificationType'
                status:
                  $ref: '#/components/schemas/NotificationStatus'
                targetCount:
                  type: integer
                  description: Number of target users
                scheduledAt:
                  type: string
                  format: date-time
                createdAt:
                  type: string
                  format: date-time
...
    DevicePlatform:
      type: string
      enum: [ios, android, web]
      description: Device platform
    NotificationType:
      type: string
      enum: [payment, booking, announcement, system, marketing, security]
      description: Type of notification
    NotificationChannel:
      type: string
      enum: [push, email, sms, inApp]
      description: Notification delivery channel
    NotificationPriority:
      type: string
      enum: [low, normal, high, urgent]
      description: Notification priority level
    NotificationStatus:
      type: string
      enum: [pending, sent, delivered, read, failed, cancelled]
      description: Notification status
  responses:
    BadRequestError:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Validation failed'
            error:
              code: 'VALIDATION_ERROR'
              details: 'token is required'
              field: 'token'
            timestamp: '2024-01-15T10:30:00Z'
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Unauthorized'
            error:
              code: 'UNAUTHORIZED'
              details: 'Invalid or missing JWT token'
            timestamp: '2024-01-15T10:30:00Z'
    ForbiddenError:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Forbidden'
            error:
              code: 'FORBIDDEN'
              details: 'Admin access required'
            timestamp: '2024-01-15T10:30:00Z'
    NotFoundError:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Not found'
            error:
              code: 'NOT_FOUND'
              details: 'Notification not found'
            timestamp: '2024-01-15T10:30:00Z'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 'Internal server error'
            error:
              code: 'INTERNAL_ERROR'
              details: 'An unexpected error occurred'
            timestamp: '2024-01-15T10:30:00Z'
...